Fix all outstanding errors and warnings during devtools::check().
This commit represents the first iteration of the package which, while the
scientific results are wildly inaccurate, is a complete refactoring and
packaging of said refactoring in a properly constructred, well-documented R
package.

There are many outstanding TODO items, and a couple of this which scream FIXME,
but overall the package and the refactoring is looking well.