Refactor the classification inhabitance, increasing performance
Built-in terra methods for (re-)classification are used to modify a copy of the
susceptible layer of the SpatRaster used throughout the main algorithm. This
improved method permits simplification of surrounding code which removed the
inhabited layer from the SpatRaster before continuing with other calculations.

The Inhabited layer is now created on-demand when it is needed.