Include test harness data to produce original code results
This includes the necessary geoTIFF and RDS file to produce a raster or vector
object used in the Q matrix-generating function from the original code (fix #10).

Tested on my machine, the originalCodeResults and the newCodeResults are both
created correctly and are now ready to be compared programmatically with the
testing framework (testthat).