Fix issue with date-time column and susceptibel count
The date-time column issue was caused by an off-by-one error (fix #5), which
also produced the issue related to the susceptible column being zero to begin
with (because the columns were right-shifted and addressed incorrectly; fix #6).
This commit fixes those two issues.

The aggregation factor used in testHMatrix is changed to reduce computational
resource demand during testing. The libraries loaded during testing have their
messaging suppressed.

The parameters used in testQMatrix are updated to use the values from the
function examples (a fix for the misusage of zero as a placeholder), and the
startup messages of libraries loaded during testing are silenced.