Improve the subassignment of global sums into summaryTable
This is fewer lines and appears to allow the simulation to run much faster, but
I don't have any scientific data (benchmarks to validate the claim); however, I
certainly wasn't able to simulate four-hundred and forty (440) days before!

Included is a check for NaN values before making a relational comparison with
numerics to prevent the error documented in #13 from occurring (fix #13).