On "rat is short for ratio!", and 1E-9; or Why Literate Programming
Ashok sent an email to Tom Bayliss White (T.B.W.) on my behalf to ask about the
choice of naming an identifier `rat` (did it have scientific meaning?) and about
the significance of number `1E-9`. I had created a comment on the BDA algorithm
implemented in our software in a shared document, reading, "What is the meaning
of "rat"? It's not the animal "rat", 🐀, so what is it? It's a ratio of the
Exposed to a proportion of the Infected... but what does it mean? Further, what
is 1E-9? Why this number? What is its significance? Is it a magic number? Can it
be anything?"

TBW responded to Ashok (forwarded to me),

"Rat is short for ratio! Also the 1E-9 is probably there in case the number of
infected people is 0 so as to not have a 0/0 error. (As if infected is 0,
exposed is likely to be 0 too). There may be an issue if Exposed were non zero
and infected zero, as that would lead to a huge increase in the number of people
down as exposed, but we haven't had that issue yet."

That is why we should use literate programming for scientific applications, or
we should at least use better identifiers and write down what is and what isn't
a "programmer's identifier" and what is a scientifically meaningful identifier
in a program.