Include SpatVector and SpatRaster as arguments to the main function
The function getSVEIRD.SpatRaster is now called for the user within the main
function, rather than the user passing in a `layers` object with these
layers (which doesn't make as much sense upon reflection). This also permits me
to make use of the original population SpatRaster or the SpatVector of
administrative boundaries in the main function (if necessary).

At one point I thought it was necessary to use the SpatVector to crop the layers
back down to their geographical area, but that was before I found the mistaken
argument list for my usage of `terra::focal` in the non-BDA algorithm.

The change to requiring the user to pass in the population SpatRaster and the
subregions SpatVector (if applicable, it is optional) is retained because I
think it is a good change.